+++ src/codec.rs
@@ -2,13 +2,16 @@ use actix_codec::{Decoder, Encoder};
 use bincode::{error::DecodeError, Decode, Encode};
 use bytes::{Buf, BufMut, BytesMut};
 use std::io;
+use std::net::IpAddr;
 
-type ConnectionID = u16;
+pub type ConnectionID = u16;
+pub type Port = u16;
 
 #[derive(Encode, Decode, Debug, PartialEq, Clone)]
 pub struct NewTCPConnection {
     pub connection_id: ConnectionID,
-    pub port: u16,
+    pub address: IpAddr,
+    pub port: Port,
 }
 
 #[derive(Encode, Decode, Debug, PartialEq, Clone)]
@@ -28,7 +31,14 @@ pub struct LogMessage {
 }
 
 #[derive(Encode, Decode, Debug, PartialEq, Clone)]
-pub enum MirrordMessage {
+pub enum ClientMessage {
+    PortSubscribe(Vec<u16>),
+    Close,
+    ConnectionUnsubscribe(Port),
+}
+
+#[derive(Encode, Decode, Debug, PartialEq, Clone)]
+pub enum DaemonMessage {
     Close,
     NewTCPConnection(NewTCPConnection),
     TCPData(TCPData),
@@ -36,26 +46,26 @@ pub enum MirrordMessage {
     LogMessage(LogMessage),
 }
 
-pub struct MirrordCodec {
+pub struct ClientCodec {
     config: bincode::config::Configuration,
 }
 
-impl MirrordCodec {
+impl ClientCodec {
     pub fn new() -> Self {
-        MirrordCodec {
+        ClientCodec {
             config: bincode::config::standard(),
         }
     }
 }
 
-impl Default for MirrordCodec {
+impl Default for ClientCodec {
     fn default() -> Self {
-        MirrordCodec::new()
+        ClientCodec::new()
     }
 }
 
-impl Decoder for MirrordCodec {
-    type Item = MirrordMessage;
+impl Decoder for ClientCodec {
+    type Item = DaemonMessage;
     type Error = io::Error;
 
     fn decode(&mut self, src: &mut BytesMut) -> io::Result<Option<Self::Item>> {
@@ -70,10 +80,61 @@ impl Decoder for MirrordCodec {
     }
 }
 
-impl Encoder<MirrordMessage> for MirrordCodec {
+impl Encoder<ClientMessage> for ClientCodec {
     type Error = io::Error;
 
-    fn encode(&mut self, msg: MirrordMessage, dst: &mut BytesMut) -> Result<(), Self::Error> {
+    fn encode(&mut self, msg: ClientMessage, dst: &mut BytesMut) -> Result<(), Self::Error> {
+        let encoded = match bincode::encode_to_vec(msg, self.config) {
+            Ok(encoded) => encoded,
+            Err(err) => {
+                return Err(io::Error::new(io::ErrorKind::Other, err.to_string()));
+            }
+        };
+        dst.reserve(encoded.len());
+        dst.put(&encoded[..]);
+
+        Ok(())
+    }
+}
+
+pub struct DaemonCodec {
+    config: bincode::config::Configuration,
+}
+
+impl DaemonCodec {
+    pub fn new() -> Self {
+        DaemonCodec {
+            config: bincode::config::standard(),
+        }
+    }
+}
+
+impl Default for DaemonCodec {
+    fn default() -> Self {
+        DaemonCodec::new()
+    }
+}
+
+impl Decoder for DaemonCodec {
+    type Item = ClientMessage;
+    type Error = io::Error;
+
+    fn decode(&mut self, src: &mut BytesMut) -> io::Result<Option<Self::Item>> {
+        match bincode::decode_from_slice(&src[..], self.config) {
+            Ok((message, read)) => {
+                src.advance(read);
+                Ok(Some(message))
+            }
+            Err(DecodeError::UnexpectedEnd) => Ok(None),
+            Err(err) => Err(io::Error::new(io::ErrorKind::Other, err.to_string())),
+        }
+    }
+}
+
+impl Encoder<DaemonMessage> for DaemonCodec {
+    type Error = io::Error;
+
+    fn encode(&mut self, msg: DaemonMessage, dst: &mut BytesMut) -> Result<(), Self::Error> {
         let encoded = match bincode::encode_to_vec(msg, self.config) {
             Ok(encoded) => encoded,
             Err(err) => {
@@ -91,27 +152,45 @@ impl Encoder<MirrordMessage> for MirrordCodec {
 mod tests {
     use super::*;
     use bytes::BytesMut;
+
     #[test]
-    fn sanity_encode_decode() {
-        let mut codec = MirrordCodec::new();
+    fn sanity_client_encode_decode() {
+        let mut client_codec = ClientCodec::new();
+        let mut daemon_codec = DaemonCodec::new();
         let mut buf = BytesMut::new();
 
-        let msg = MirrordMessage::NewTCPConnection(NewTCPConnection {
-            connection_id: 1,
-            port: 8080,
-        });
+        let msg = ClientMessage::PortSubscribe(vec![1, 2, 3]);
 
-        codec.encode(msg.clone(), &mut buf).unwrap();
+        client_codec.encode(msg.clone(), &mut buf).unwrap();
 
-        let decoded = codec.decode(&mut buf).unwrap().unwrap();
+        let decoded = daemon_codec.decode(&mut buf).unwrap().unwrap();
 
         assert_eq!(decoded, msg);
         assert!(buf.is_empty());
     }
 
     #[test]
-    fn decode_invalid_data() {
-        let mut codec = MirrordCodec::new();
+    fn sanity_daemon_encode_decode() {
+        let mut client_codec = ClientCodec::new();
+        let mut daemon_codec = DaemonCodec::new();
+        let mut buf = BytesMut::new();
+
+        let msg = DaemonMessage::TCPData(TCPData {
+            connection_id: 1,
+            data: vec![1, 2, 3],
+        });
+
+        daemon_codec.encode(msg.clone(), &mut buf).unwrap();
+
+        let decoded = client_codec.decode(&mut buf).unwrap().unwrap();
+
+        assert_eq!(decoded, msg);
+        assert!(buf.is_empty());
+    }
+
+    #[test]
+    fn decode_client_invalid_data() {
+        let mut codec = ClientCodec::new();
         let mut buf = BytesMut::new();
         buf.put_u8(254);
 
@@ -123,10 +202,32 @@ mod tests {
     }
 
     #[test]
-    fn decode_partial_data() {
-        let mut codec = MirrordCodec::new();
+    fn decode_client_partial_data() {
+        let mut codec = ClientCodec::new();
         let mut buf = BytesMut::new();
         buf.put_u8(1);
+
+        assert!(codec.decode(&mut buf).unwrap().is_none());
+    }
+
+    #[test]
+    fn decode_daemon_invalid_data() {
+        let mut codec = DaemonCodec::new();
+        let mut buf = BytesMut::new();
+        buf.put_u8(254);
+
+        let res = codec.decode(&mut buf);
+        match res {
+            Ok(_) => panic!("Should have failed"),
+            Err(err) => assert_eq!(err.kind(), io::ErrorKind::Other),
+        }
+    }
+
+    #[test]
+    fn decode_daemon_partial_data() {
+        let mut codec = DaemonCodec::new();
+        let mut buf = BytesMut::new();
+        buf.put_u8(0);
 
         assert!(codec.decode(&mut buf).unwrap().is_none());
     }