+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 ## [Unreleased]
+## 0.1.0 - 25/04/2022
+### Changed
+- Split messages into Client messages and Daemon messages.
+- There's now codec for each side.
+- Added new message types
 
 ## 0.0.1 - 14/3/2022
 Initial release